Robo Grazer Schematics
The schematics are used with boards that use wire wrapping. This is very flexible during development and seems to have been quite reliable.
The boards have a backplane with a  24 way connector. Each have the components mounted on veroboard with a 24way front connector to carry external cables.
The only board that is separated is the motor driver which is mounted as near the motors as possible.